historical_processor/requirements.txt
# Historical Episodes Processor - Dependencies
# Python 3.8+

# Core dependencies
openai>=1.0.0
elevenlabs>=2.0.0
pyyaml>=6.0
colorama>=0.4.6
langdetect>=1.0.9

# Optional for development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.3.0
pylint>=2.17.0
mypy>=1.4.0
